requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
openai==1.58.1
azure-cosmos==4.9.0
azure-search-documents==11.6.0b8
azure-storage-blob==12.24.0
azure-identity==1.19.0
structlog==24.4.0
azure-ai-documentintelligence==1.0.0
httpx==0.28.1
scikit-learn==1.6.1
pytest==8.3.4
pytest-asyncio==0.25.0
